+++ apps/stats/src/views/Stats/Growth.tsx
@@ -1,154 +1,24 @@
 import CustomTooltipContent from '@src/components/chart/CustomTooltipContent';
 import DateRangeSelect from './components/DateRangeSelect';
-import React, {useMemo, useState} from 'react';
+import React, {useState} from 'react';
 import StatsLayout from './layout/StatsLayout';
 import StatsView from './layout/StatsView';
-import moment from 'moment';
 import {Card, CardContent, CardDescription, CardHeader, CardTitle, ChartConfig, ChartContainer, ChartTooltip, H1, Recharts, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, Tabs, TabsList, ViewHeader, ViewHeaderActions, formatDisplayDate, formatNumber} from '@tryghost/shade';
 import {KpiTabTrigger, KpiTabValue} from './components/KpiTab';
-import {MemberStatusItem, useMemberCountHistory} from '@tryghost/admin-x-framework/api/stats';
 import {Navigate} from '@tryghost/admin-x-framework';
 import {calculateYAxisWidth, getYTicks} from '@src/utils/chart-helpers';
 import {getSettingValue} from '@tryghost/admin-x-framework/api/settings';
 import {useGlobalData} from '@src/providers/GlobalDataProvider';
+import {useGrowthStats} from '@src/hooks/useGrowthStats';
 
-// Type for direction values
-type DiffDirection = 'up' | 'down' | 'same';
-
-// Helper function to convert range to date parameters
-const getRangeDates = (rangeInDays: number) => {
-    const endDate = moment().format('YYYY-MM-DD');
-    let startDate;
-    
-    if (rangeInDays === 1) {
-        // Today
-        startDate = endDate;
-    } else if (rangeInDays === 1000) {
-        // All time - use a far past date
-        startDate = '2010-01-01';
-    } else {
-        // Specific range
-        startDate = moment().subtract(rangeInDays - 1, 'days').format('YYYY-MM-DD');
-    }
-    
-    return {startDate, endDate};
-};
-
-// Extract the calculation to a separate function to avoid conditional hook calls
-const calculateTotals = (memberData: MemberStatusItem[]) => {
-    if (!memberData.length) {
-        return {
-            totalMembers: 0,
-            freeMembers: 0,
-            paidMembers: 0,
-            percentChanges: {
-                total: '0%',
-                free: '0%',
-                paid: '0%'
-            },
-            directions: {
-                total: 'same' as DiffDirection,
-                free: 'same' as DiffDirection,
-                paid: 'same' as DiffDirection
-            }
-        };
-    }
-    
-    // Get latest values
-    const latest = memberData[memberData.length - 1];
-    
-    // Calculate total members
-    const totalMembers = latest.free + latest.paid + latest.comped;
-    
-    // Calculate percentage changes if we have enough data
-    const percentChanges = {
-        total: '0%',
-        free: '0%',
-        paid: '0%'
-    };
-    
-    const directions = {
-        total: 'same' as DiffDirection,
-        free: 'same' as DiffDirection,
-        paid: 'same' as DiffDirection
-    };
-    
-    if (memberData.length > 1) {
-        // Get first day in range
-        const first = memberData[0];
-        const firstTotal = first.free + first.paid + first.comped;
-        
-        if (firstTotal > 0) {
-            const totalChange = ((totalMembers - firstTotal) / firstTotal) * 100;
-            percentChanges.total = `${Math.abs(totalChange).toFixed(1)}%`;
-            directions.total = totalChange > 0 ? 'up' : totalChange < 0 ? 'down' : 'same';
-        }
-        
-        if (first.free > 0) {
-            const freeChange = ((latest.free - first.free) / first.free) * 100;
-            percentChanges.free = `${Math.abs(freeChange).toFixed(1)}%`;
-            directions.free = freeChange > 0 ? 'up' : freeChange < 0 ? 'down' : 'same';
-        }
-        
-        if (first.paid > 0) {
-            const paidChange = ((latest.paid - first.paid) / first.paid) * 100;
-            percentChanges.paid = `${Math.abs(paidChange).toFixed(1)}%`;
-            directions.paid = paidChange > 0 ? 'up' : paidChange < 0 ? 'down' : 'same';
-        }
-    }
-    
-    return {
-        totalMembers,
-        freeMembers: latest.free,
-        paidMembers: latest.paid,
-        percentChanges,
-        directions
-    };
-};
-
-// Format chart data outside component
-const formatChartData = (memberData: MemberStatusItem[]) => {
-    return memberData.map(item => ({
-        date: item.date,
-        value: item.free + item.paid + item.comped,
-        formattedValue: formatNumber(item.free + item.paid + item.comped),
-        label: 'Total members'
-    }));
-};
-
-const GrowthKPIs:React.FC = () => {
+const GrowthKPIs: React.FC = () => {
     const [currentTab, setCurrentTab] = useState('free-members');
     const {settings, range} = useGlobalData();
     const labs = JSON.parse(getSettingValue<string>(settings, 'labs') || '{}');
 
-    // Get date range using useMemo to prevent unnecessary recalculations
-    const {startDate, endDate} = useMemo(() => getRangeDates(range), [range]);
-
-    // Fetch member count history from API
-    const {data: memberCountResponse} = useMemberCountHistory({
-        searchParams: {
-            date_from: startDate,
-            date_to: endDate
-        }
-    });
-    
-    // Wrap memberData in useMemo to maintain stable reference
-    const memberData = useMemo(() => memberCountResponse?.data || [], [memberCountResponse?.data]);
-    
-    // Use useMemo to calculate totals
-    const totalsData = useMemo(() => calculateTotals(memberData), [memberData]);
-    
-    // Destructure the totals
-    const {
-        totalMembers,
-        freeMembers,
-        paidMembers,
-        percentChanges,
-        directions
-    } = totalsData;
-
-    // Chart data
-    const chartData = useMemo(() => formatChartData(memberData), [memberData]);
+    // Get growth stats from custom hook
+    const {chartData, totals} = useGrowthStats(range);
+    const {totalMembers, freeMembers, paidMembers, percentChanges, directions} = totals;
 
     if (!labs.trafficAnalyticsAlpha) {
         return <Navigate to='/' />;
@@ -258,10 +128,11 @@ const GrowthKPIs:React.FC = () => {
     );
 };
 
-const Growth:React.FC = () => {
+const Growth: React.FC = () => {
     const {range} = useGlobalData();
-    // Use useMemo to prevent unnecessary recalculations
-    const {startDate, endDate} = useMemo(() => getRangeDates(range), [range]);
+    
+    // Get stats from custom hook
+    const {isLoading} = useGrowthStats(range);
     
     // TODO: Replace this with real top posts data from API
     const mockTopPosts = [
@@ -281,14 +152,6 @@ const Growth:React.FC = () => {
         {id: 'post-014', title: 'Creating a Productive Workspace on a Budget', freeMembers: 1, paidMembers: 0, mrr: 0},
         {id: 'post-015', title: 'Simple Ways to Reduce Your Carbon Footprint', freeMembers: 1, paidMembers: 0, mrr: 0}
     ];
-
-    // Get member count data to determine if we're loading
-    const {isLoading} = useMemberCountHistory({
-        searchParams: {
-            date_from: startDate,
-            date_to: endDate
-        }
-    });
 
     return (
         <StatsLayout>
